+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal 0.15.2.9046
+# teal 0.15.2.9047
 
 ### Enhancement
 * Provided progress bar for modules loading and data filtering during teal app startup.